Moved test_spy.py inside test_image.py
